Challenge 1B/requirements.txt
pdfminer.six==20221105
pdfplumber==0.7.6
scikit-learn==1.2.2
numpy==1.24.2